+++ demo-app/config/routes.rb
@@ -21,6 +21,12 @@ Rails.application.routes.draw do
     resources :posts
     resource :hide_from_backend, only: [:new, :create]
     resources :smooth_redirects, only: [:new, :create]
+    resources :races, only: [:index] do
+      collection do
+        get :morph_preserves_dialog
+        get :supersede_close
+      end
+    end
     root to: "welcome#index"
   end